requirements.txt
python-dotenv
openai
llama-index
streamlit
nltk

